Added search books functionality
Within the rent section of the app, the searching allows you to search for your desired book. IMPORTANT --> it shows recommendations only if a book is not found, will come back and change this to show recommendations even if a valid book is entered. Also, recommendations aren't gonna show up in the panel for rec's --> look nicer imo