done req9
made there be more space for the message in the notifications panel.
FYI - the whole idea is that it's reading the 'CourseTextbooks.csv' file to check if it should be high priority or not. this way we can determine if it should be high/low priority, as if it is contained in that csv file, it is a course teaching txtbook which is the parameter for a high priority